Clean up: remove old code & do some organization
